// SFA_MOBILE_API/src/modules/mobile/transactions/types/sales.types.ts
export type InvoiceHeaderUploadItem = {
  transactionkey: string | number;
  routekey: string | number;
  visitkey: string | number;
  documentnumber: string | number;
  invoicenumber: string | number;
  transactiondate: string;
  transactiontime: string;
  dsdnumber: string;
  ponumber: string;
  customercode: string | number;
  routecode: string | number;
  salesmancode: string | number;
  presoldordernumber: string | number;
  presalesmancode: string | number;
  presalesroutecode: string | number;
  orderdeliverydate: string;
  orderdeliveryroutecode: string | number;
  totalinvoiceamount: string | number;
  totalsalesamount: string | number;
  totalreturnamount: string | number;
  totaldamagedamount: string | number;
  totalfreesampleamount: string | number;
  immediatepaid: string | number;
  amountpaid: string | number;
  invoicebalance: string | number;
  dexflag: string | number;
  dexg86signature: string;
  paymenttype: string | number;
  splittransaction: string | number;
  voidflag: string | number;
  transmitindicator: string | number;
  paymentstatus: string | number;
  hhcinvoicenumber: string | number;
  totalpromoamount: string | number;
  gcpaymenttype: string | number;
  hhcdocumentnumber: string | number;
  inventorykey: string | number;
  totaltaxesamount: string | number;
  itemlinetaxamount: string | number;
  totaldiscountamount: string | number;
  voidreasoncode: string | number;
  totalexpiryamount: string | number;
  currencycode: string | number;
  totalmanualfree: string | number;
  totallimitedfree: string | number;
  totalrebaterent: string | number;
  totalfixedrent: string | number;
  actualtransactiondate: string;
  boentry: string | number;
  hhctransactionkey: string | number;
  data: string | number;
  comments: string;
  totaldiscdistributionamount: string | number;
  totalreplacementamount: string | number;
  comments2: string;
  lineitemdiscount: string | number;
  totalreturnpromoamount: string | number;
  returnlineitemdiscount: string | number;
  totalbuybackfreeamount: string | number;
  roundtotalsalesamount: string | number;
  diffround: string | number;
  headerdiscounttaxamt: string | number;
};

export type InvoiceDetailUploadItem = {
  routekey: string | number;
  visitkey: string | number;
  transactionkey: string | number;
  itemcode: string | number;
  salesqty: string | number;
  returnqty: string | number;
  damagedqty: string | number;
  freesampleqty: string | number;
  salesprice: string | number;
  returnprice: string | number;
  stdsalesprice: string | number;
  stdreturnprice: string | number;
  promoqty: string | number;
  salesitemexcisetax: string | number;
  salesitemgsttax: string | number;
  returnitemexcisetax: string | number;
  returnitemgsttax: string | number;
  damageditemexcisetax: string | number;
  damageditemgsttax: string | number;
  fgitemexcisetax: string | number;
  fgitemgsttax: string | number;
  promoitemexcisetax: string | number;
  promoitemgsttax: string | number;
  coopid: string;
  batchdetailkey: string | number;
  salescaseprice: string | number;
  returncaseprice: string | number;
  stdsalescaseprice: string | number;
  stdreturncaseprice: string | number;
  goodreturnprice: string | number;
  goodreturncaseprice: string | number;
  stdgoodreturncaseprice: string | number;
  stdgoodreturnprice: string | number;
  expiryqty: string | number;
  currencycode: string | number;
  returnfreeqty: string | number;
  manualfreeqty: string | number;
  limitedfreeqty: string | number;
  rebaterentqty: string | number;
  fixedrentqty: string | number;
  pricechgindicator: string | number;
  discountamount: string | number;
  discountpercentage: string | number;
  promoamount: string | number;
  replacementqty: string | number;
  replacementprice: string | number;
  replacementcaseprice: string | number;
  promovalue: string | number;
  mdat: string;
  sales_amount: string | number;
  returnpromovalue: string | number;
  returnpromoamount: string | number;
  return_amount: string | number;
  returnfreesampleqty: string | number;
  roundsalesamount: string | number;
  diffround: string | number;
  amount: string | number;
};

export type InvoiceRxdDetailUploadItem = {
  routekey: string | number;
  visitkey: string | number;
  transactionkey: string | number;
  itemtransactiontype: string | number;
  itemcode: string | number;
  itemtranstypeseq: string | number;
  reasoncode: string | number;
  quantity: string | number;
  catchweightqty: string | number;
  weighted: string | number;
  instructioncode: string | number;
  currencycode: string | number;
  expirydate: string;
  invoicenumber: string | number;
};

export type PromotionDetailUploadItem = {
  routekey: string | number;
  visitkey: string | number;
  transactionkey: string | number;
  itemtransactiontype: string | number;
  itemcode: string | number;
  promotiontypecode: string | number;
  promotionamount: string | number;
  promotionquantity: string | number;
  catchweightqty: string | number;
  weighted: string | number;
  promotionplannumber: string | number;
  assignmentkey: string | number;
  exclusionoption: string | number;
  promochgindicator: string | number;
  oldpromotionamount: string | number;
  performindicator: string | number;
  performcriteriakey: string | number;
  promotioncaseprice: string | number;
  currencycode: string | number;
  memo1: string;
};

export type CustomerInvoiceUploadItem = {
  transactionkey: string | number;
  transactiontype: string | number;
  documentnumber: string | number;
  invoicenumber: string | number;
  transactiondate: string;
  transactiontime: string;
  customercode: string | number;
  routecode: string | number;
  salesmancode: string | number;
  totalinvoiceamount: string | number;
  totalsalesamount: string | number;
  totalreturnamount: string | number;
  totaldamagedamount: string | number;
  totalfreesampleamount: string | number;
  immediatepaid: string | number;
  amountpaid: string | number;
  dnamountpaid: string | number;
  cnamountpaid: string | number;
  invoicebalance: string | number;
  paymenttype: string | number;
  voidflag: string | number;
  paymentstatus: string | number;
  hhcinvoicenumber: string | number;
  remarks1: string;
  remarks2: string;
  routestartdate: string;
  erpreferencenumber: string;
  mdat: string;
  totalpromoamount: string | number;
  gcpaymenttype: string | number;
  totaltaxesamount: string | number;
  itemlinetaxamount: string | number;
  totaldiscountamount: string | number;
  pdcindicator: string | number;
  chequecollection: string | number;
  totalexpiryamount: string | number;
  currencycode: string | number;
  pdcbalance: string | number;
  totalmanualfree: string | number;
  totallimitedfree: string | number;
  totalrebaterent: string | number;
  totalfixedrent: string | number;
  data: string | number;
  totaldiscdistributionamount: string | number;
  totalreplacementamount: string | number;
  pdcdate: string;
  totalbuybackfreeamount: string | number;
  duedate: string;
};

export type BatchExpiryDetailUploadItem = LegacyUploadItem;